small tweaks, design, usability/accessibility
