+++ components/footer.php
@@ -1,6 +1,4 @@
-
- <!-- Footer -->
- <section class="pt-5">
+<section class="pt-5">
       <!-- Whitespace on top -->
   </section>
     
@@ -12,9 +10,7 @@
   <!-- Section: Links  -->
   <section class="">
     <div class="container text-center text-md-start mt-5">
-      <!-- Grid row -->
       <div class="row mt-3">
-        <!-- Grid column -->
         <div class="col-md-3 col-lg-4 col-xl-3 mx-auto mb-4">
         <!-- Content -->
           <h6 class="text-uppercase fw-bold mb-4">
@@ -24,11 +20,6 @@
             Sapphire Shores Hotel
           </p>
         </div>
-        <!-- Grid column -->
-
-        <!-- Grid column -->
-
-        <!-- Grid column -->
         <div class="col-md-3 col-lg-2 col-xl-2 mx-auto mb-4">
           <!-- Links -->
           <h6 class="text-uppercase fw-bold mb-4">
@@ -42,9 +33,6 @@
           </p>
 
         </div>
-        <!-- Grid column -->
-
-        <!-- Grid column -->
         <div class="col-md-4 col-lg-3 col-xl-3 mx-auto mb-md-0 mb-4">
           <!-- Links -->
           <h6 class="text-uppercase fw-bold mb-4">Contact</h6>
@@ -57,9 +45,7 @@
             wi22b076@technikum-wien.at
           </p>
         </div>
-        <!-- Grid column -->
       </div>
-      <!-- Grid row -->
     </div>
   </section>
   <!-- Copyright -->
