+++ details.php
@@ -42,11 +42,9 @@ require_once("data/dbaccess.php");
 </head>
 
 <body>
-    <header>
-        <!-- header -->
-    </header>
-
-    <?php include 'components/navbar.php'; ?>
+    <nav>
+    <?php include 'components/navbar.php'; ?> 
+    </nav>
     <?php
     // check if user is logged in
     if (!isset($_SESSION['loggedIn']) || $_SESSION['loggedIn'] !== true) {
@@ -108,8 +106,9 @@ require_once("data/dbaccess.php");
         }
         ?>
     </main>
-
-    <?php include 'components/footer.php'; ?>
+    <footer>
+       <?php include 'components/footer.php'; ?> 
+    </footer>
     <?php include 'includes/scripts.php'; ?>
 </body>
