+++ faq.php
@@ -1,3 +1,7 @@
+<?php
+require_once("data/dbaccess.php");
+?>
+
 <!DOCTYPE html>
 <html lang="de">
 <head>
@@ -6,11 +10,9 @@
     <title>FAQ</title>
 </head>
 <body>
-    <header>
-        <!-- Header content -->
-    </header>
-
-    <?php include 'components/navbar.php'; ?>
+    <nav>
+       <?php include 'components/navbar.php'; ?> 
+    </nav>
 
     <main class="container my-4">
         <div class="row">
@@ -103,8 +105,9 @@
             </div>
         </div>
     </main>
-
-    <?php include 'components/footer.php'; ?>
+    <footer>
+        <?php include 'components/footer.php'; ?>
+    </footer>
     <?php include 'includes/scripts.php'; ?>
     <!-- Bootstrap JS -->
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
