+++ login.php
@@ -17,11 +17,9 @@
     </style>
 </head>
 <body>
-    <header>
-        <!-- Header content -->
-    </header>
-
-    <?php include 'components/navbar.php'; ?>
+    <nav>
+        <?php include 'components/navbar.php'; ?>
+    </nav>
 
     <main>
         <div class="container mt-3 pt-5">
@@ -73,8 +71,9 @@
             </form>
         </div>
     </main>
-
-    <?php include 'components/footer.php'; ?>
+    <footer>
+       <?php include 'components/footer.php'; ?> 
+    </footer>
     <?php include 'includes/scripts.php'; ?>
 </body>
 </html>
